Add new input type: memory image
A memory image is similar to an object file, but each line in the
memory image file contains an address and the full word of data
at that address. No comments and no blank lines are allowed.
For now, memory must start at address 0, and data must be
contiguous (can't skip and memory addresses yet).